requirements.txt
streamlit>=1.20.0
pillow
pytesseract
# easyocr requires torch; install only if you plan to use it:
easyocr
